third_party/WebKit/LayoutTests/fast/css/usecounter-negative-mask-size-longhand.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
'use strict';

test(() => {
    let NegativeMaskSize = 2403; // From web_feature.mojom

    let isCounted = () => internals.isUseCounted(document, NegativeMaskSize);
    var div = document.createElement('div');

    div.style = '-webkit-mask-size: 1px 2px;';
    div.style = 'background-size: -1px -2px;';
    assert_false(isCounted(),
                 'background-size should not be counted');

    div.style = '-webkit-mask-size: 1px -2px;';
    assert_true(isCounted(),
                '-webkit-mask-size should be counted');
}, 'Negative size is use counted for -webkit-mask-size');
</script>
